requirements.txt
streamlit
pandas
matplotlib
seaborn
pyxlsb
groq
python-dotenv
